fix(gemma4): keep attached MTP API staged
