+++ src/llama-ext.h
@@ -90,6 +90,32 @@ LLAMA_API ggml_backend_dev_t llama_model_get_device(const struct llama_model * m
 LLAMA_API llama_memory_breakdown llama_get_memory_breakdown(const struct llama_context * ctx);
 
 //
+// attached MTP assistant decoding
+//
+
+// Load an attached MTP assistant into a compatible target before creating the context.
+LLAMA_API int llama_model_load_mtp_from_file(
+        struct llama_model * model,
+        const char * path_mtp,
+        struct llama_model_params params);
+
+LLAMA_API const struct llama_model * llama_model_get_mtp_assistant(const struct llama_model * model);
+LLAMA_API bool llama_model_has_mtp_assistant(const struct llama_model * model);
+LLAMA_API uint32_t llama_model_mtp_n_embd_backbone(const struct llama_model * model);
+
+// Greedy draft using the target context KV and a loaded attached MTP assistant.
+LLAMA_API int32_t llama_decode_mtp(
+        struct llama_context * ctx,
+        llama_seq_id seq_id,
+        llama_pos attn_pos,
+        llama_token last_token,
+        float * h_prev,
+        int32_t n_steps,
+        llama_token * out_drafts,
+        float * out_logits,
+        float * out_h_prev_last);
+
+//
 // pre-norm embeddings (hidden state before the final output norm)
 //
